Readd test for obsolete ToQuery
Fixes #24623
