+++ src/EFCore/Design/Internal/CSharpRuntimeAnnotationCodeGenerator.cs
@@ -64,11 +64,15 @@ namespace Microsoft.EntityFrameworkCore.Design.Internal
             var annotations = parameters.Annotations;
             if (!parameters.IsRuntime)
             {
-                annotations.Remove(CoreAnnotationNames.PropertyAccessMode);
-                annotations.Remove(CoreAnnotationNames.NavigationAccessMode);
-                annotations.Remove(CoreAnnotationNames.DiscriminatorProperty);
-                annotations.Remove(CoreAnnotationNames.AmbiguousNavigations);
-                annotations.Remove(CoreAnnotationNames.NavigationCandidates);
+                foreach (var annotation in annotations)
+                {
+                    if (CoreAnnotationNames.AllNames.Contains(annotation.Key)
+                        && annotation.Key != CoreAnnotationNames.DiscriminatorValue
+                        && annotation.Key != CoreAnnotationNames.DiscriminatorMappingComplete)
+                    {
+                        annotations.Remove(annotation.Key);
+                    }
+                }
             }
 
             GenerateSimpleAnnotations(parameters);
